Move FetchChargeVersions
